fix favicon was not shown properly
